a lot of things got done
so, i added a lot, and created many errors.
so many it took a week to resolve them all.

also i went on vacation for a little over a week in the middle of this
AND broke my left arm so its been a slow journey for me

either way enjoy the million things i have been working for the past month

bye