it works
ok, so not every type is done yet, but the basic structure is all done.
You can feed the game an XML file,
and it will extract most of the items out of it.

The game does not yet run on this new system,
but it should be a smooth transition.

Some features to come:

All types convertible from XML,
Game will run using the XML data,
More configuration will be stored in XML,
Lots of typos will be fixed,
...and more!

Unfortunately there is no visible difference in the game because of these changes,
but they will make developement faster and more accessible.

Please enjoy!
-Foxler2010