tidy up everything.
here is a laundry list of all the changes i made:

- fixed the coolTrophy item having a type of 8 instead of 9
- finished updating the enum values to reflect the new chnages in ItemType
- Renamed the Other class to Miscellaneous. Please tell me if I spelled it wrong.
- Created a temporary ItemDataReader in Data because I needed to use it inside a construct() method
- Created the Trophy class, which is identical to Junk. Now the Trophy is entirely separate from Junk
- Added indexes to Options, similar to ItemType's indexes.
These are used in the new Option counting system
- The new Option Counting System (or OCS, for short),
puts all the visit counters into an ArrayList and uses Option indexes to track which is which.
Inside opCode(), visitUp() is used to add visits.
In Main, getVisits() along with Option's getIndex() is used to get the visits for the current option automatically.
- Probably more stuff I haven't listed it all.

Enjoy this new addition!
-Foxler2010